+++ HISTORY.md
@@ -3,4 +3,4 @@
 
 ## 1.0.0 / 2016-11-29
 
--
+- inspired by react-native